chore: folder structure refinements
