+++ lib/src/constants/app_text.dart
@@ -1,0 +1,3 @@
+class AppText {
+  static const tapValidationText = 'taps must be greater than 0';
+}
