+++ lib/src/constants/enums.dart
@@ -1,0 +1,6 @@
+enum HapticFeedbackType {
+  lightImpact,
+  mediumImpact,
+  heavyImpact,
+  selectionClick,
+}
