+++ lib/src/models/haptic_feed_back_types.dart
@@ -1,0 +1,6 @@
+enum HapticFeedbackType {
+  lightImpact,
+  mediumImpact,
+  heavyImpact,
+  selectionClick,
+}
